+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 ### 12.120.0
 
-- `[knf]` Added `Alias` method
+- `[knf]` Added methods `Alias` and `Config.Alias`
 - `[knf]` Added property name validation for all getters
+- `[sliceutil]` Added method `IsEqual`
 - `[knf]` Code refactoring
 - `[knf]` Added more tests